rdma_monitor/requirements.txt
prometheus-client>=0.20.0
pyyaml>=6.0
requests>=2.31.0
